requirements.txt
torch
torchvision
numpy
imageio
scipy
Pillow
scikit-image
matplotlib
streamlit